Move to AbstractGroovyStubSourcesMojo
